Improve episode release time information
